Client/requirements.txt
# OS Is included natively, otherwise you may need to pip install that too!
requests
websocket-client
